+++ NEWS.md
@@ -1,8 +1,8 @@
-# mirai 0.1.0.9003 (development)
+# mirai 0.1.0.9004 (development)
 
-* The evaluated result of a mirai is now stored at `$data` rather than `$value`. This may be queried at any time and returns an NA 'unresolved value' if the async operation is yet to complete.
+* The evaluated result of a mirai is now stored at `$data` and may be accessed at any time, returning an NA 'unresolved value' if the async operation is yet to complete. Use `call_mirai()` only to call and wait for  completion of an Aio.
 * `eval_mirai()` adds support for evaluating arbitrary length expressions wrapped in `{}`.
-* Auxiliary functions `is_nul_byte()` and `is_resolved()` re-exported from the nanonext package.
+* Auxiliary functions `is_nul_byte()` and `unresolved()` re-exported from the nanonext package.
 * New `mirai()` daemon manager for persistent background processes [experimental].
 
 # mirai 0.1.0